Remove ECSlidingUnwindSegue
This class was only used internally by ECSlidingViewController, but the
header and symbol is still available to users. Also, "slidingUnwind"
shows up in the list of segues when making connections in Xcode 5.

The solution was to remove it completely and to use ECSlidingSegue handle
unwinding by adding a private BOOL that can be set by
ECSlidingViewController.